requirements.txt
absl-py==2.2.2
actionlib==1.14.3
angles==1.9.14
apt-clone==0.2.1
apturl==0.5.2
argcomplete==1.8.1
astunparse==1.6.3
asyncio==3.4.3
atomicwrites==1.1.5
attrs==19.3.0
autobahn==17.10.1
Automat==0.8.0
bcrypt==3.1.7
beautifulsoup4==4.12.2
blinker==1.4
bondpy==1.8.7
breezy==3.0.2
Brlapi==0.7.0
cachetools==5.5.2
cam==0.3.1
camera-calibration==1.17.0
camera-calibration-parsers==1.12.1
catkin==0.8.12
catkin-pkg==1.0.0
catkin-pkg-modules==1.0.0
catkin-tools==0.9.4
cbor==1.0.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==3.3.2
Click==7.0
colcon-argcomplete==0.3.3
colcon-bash==0.5.0
colcon-cd==0.2.1
colcon-cmake==0.2.29
colcon-common-extensions==0.3.0
colcon-core==0.19.0
colcon-defaults==0.2.9
colcon-devtools==0.3.0
colcon-installed-package-information==0.2.1
colcon-library-path==0.2.1
colcon-metadata==0.2.5
colcon-notification==0.3.0
colcon-output==0.2.13
colcon-override-check==0.0.1
colcon-package-information==0.4.0
colcon-package-selection==0.2.10
colcon-parallel-executor==0.3.0
colcon-pkg-config==0.1.0
colcon-powershell==0.4.0
colcon-python-setup-py==0.2.9
colcon-recursive-crawl==0.2.3
colcon-ros==0.5.0
colcon-test-result==0.3.8
colcon-zsh==0.5.0
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
controller-manager==0.20.0
controller-manager-msgs==0.20.0
cov-core==1.15.0
coverage==4.5.2
cpplint==1.6.1
cryptography==2.8
cupshelpers==1.0
cv-bridge==1.16.2
cycler==0.10.0
Cython==3.0.5
dataclasses==0.6
dbus-python==1.2.16
decorator==4.4.2
defer==1.0.6
defusedxml==0.6.0
Deprecated==1.2.7
diagnostic-analysis==1.12.1
diagnostic-common-diagnostics==1.12.1
diagnostic-updater==1.12.1
distlib==0.3.0
distro==1.4.0
distro-info==0.23+ubuntu1.1
docutils==0.16
dulwich==0.19.15
duplicity==0.8.12.0
dynamic-reconfigure==1.7.5
empy==3.3.2
entrypoints==0.3
fasteners==0.14.1
fastimport==0.9.8
filelock==3.13.1
flake8==3.7.9
flatbuffers==25.2.10
future==0.18.3
gast==0.4.0
gazebo_plugins==2.9.3
gazebo_ros==2.9.3
gdown==4.7.1
gencpp==0.7.2
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.1
gennodejs==2.0.2
genpy==0.6.18
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.40.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
gpg==1.13.1
graphsurgeon==0.4.6
grpcio==1.70.0
h5py==3.11.0
html5lib==1.0.1
httplib2==0.14.0
hyperlink==19.0.0
idna==2.8
ifcfg==0.18
image-geometry==1.16.2
imageio==2.32.0
importlib_metadata==8.5.0
incremental==16.10.1
interactive-markers==1.12.2
Jetson.GPIO==2.0.20
joint-state-publisher==1.15.2
joint-state-publisher-gui==1.15.2
keras==2.13.1
keyboard==0.13.5
keyring==18.0.1
kiwisolver==1.0.1
language-selector==0.1
lark-parser==0.8.1
laser-geometry==1.6.8
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy_loader==0.3
libclang==18.1.1
lidar==0.7.2
lockfile==0.12.2
loguru==0.7.3
louis==3.12.0
lxml==4.5.0
lz4==3.0.2+dfsg
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.3.4
mccabe==0.6.1
message-filters==1.17.3
mock==3.0.5
monotonic==1.5
more-itertools==4.2.0
mpi4py==3.0.3
netifaces==0.10.4
networkx==3.1
ninja==1.11.1.4
nose==1.3.7
nose2==0.9.1
notify2==0.3
numpy==1.24.4
oauthlib==3.1.0
olefile==0.46
onboard==1.4.1
onnx==1.8.1
onnx-simplifier==0.3.5
onnxoptimizer==0.3.13
onnxruntime==1.8.0
opencv-python==4.12.0.88
opt_einsum==3.4.0
osrf-pycommon==2.0.2
packaging==23.2
PAM==0.4.2
pandas==2.0.3
paramiko==2.6.0
pbr==5.4.5
pexpect==4.6.0
pillow==10.4.0
pluggy==0.13.0
polars==1.8.2
protobuf==3.20.3
psutil==5.9.6
py==1.8.1
py-cpuinfo==9.0.0
py-ubjson==0.14.0
pyaml==23.12.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycocotools==2.0.7
pycodestyle==2.5.0
pycrypto==2.6.1
pycryptodomex==3.6.1
pycups==1.9.73
pydocstyle==2.1.1
pydot==1.4.1
pyflakes==2.1.1
PyGithub==1.43.7
Pygments==2.3.1
PyGObject==3.36.0
PyHamcrest==1.9.0
PyICU==2.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
PyOpenGL==3.1.0
pyOpenSSL==19.0.0
pyparsing==2.4.6
pypng==0.0.20
PyQRCode==1.2.1
PyQt5==5.14.1
pyRFC3339==1.1
pyserial==3.5
pyshp==2.3.1
PySimpleGUI==2.30.0
PySocks==1.7.1
pytesseract==0.3.13
pytest==4.6.9
pytest-cov==2.8.1
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-dbusmock==0.19
python-debian==0.1.36+ubuntu1.1
python-gitlab==2.0.1
python-gnupg==0.4.5
python-qt-binding==0.4.6
python-snappy==0.5.3
PyTrie==0.2
pytz==2023.3.post1
PyWavelets==1.4.1
pyxdg==0.26
PyYAML==6.0.1
pyzed @ file:///tmp/selfgz21785/pyzed-4.0-cp38-cp38-linux_aarch64.whl#sha256=c38de796300a5d47ed6b5195018c02c5281e95644630fb7657f675bd3319336a
qt-dotgraph==0.4.5
qt-gui==0.4.5
qt-gui-cpp==0.4.5
qt-gui-py-common==0.4.5
requests==2.32.4
requests-oauthlib==2.0.0
requests-unixsocket==0.2.0
resource-retriever==1.12.10
richdem==0.3.4
roman==2.0.0
rosbag==1.17.3
rosboost-cfg==1.15.10
rosclean==1.15.10
roscreate==1.15.10
rosdep==0.25.1
rosdep-modules==0.25.1
rosdistro-modules==1.0.1
rosgraph==1.17.3
rosinstall-generator==0.1.23
roslaunch==1.17.3
roslib==1.15.10
roslint==0.12.0
roslz4==1.17.3
rosmake==1.15.10
rosmaster==1.17.3
rosmsg==1.17.3
rosnode==1.17.3
rosparam==1.17.3
rospkg==1.6.0
rospkg-modules==1.6.0
rospy==1.17.3
rosservice==1.17.3
rostest==1.17.3
rostopic==1.17.3
rosunit==1.15.10
roswtf==1.17.3
rplidar-roboticia==0.9.5
rqt-action==0.4.11
rqt-bag==0.5.3
rqt-bag-plugins==0.5.3
rqt-console==0.4.14
rqt-dep==0.4.14
rqt-graph==0.4.16
rqt-gui==0.5.5
rqt-gui-py==0.5.5
rqt-image-view==0.4.19
rqt-launch==0.4.10
rqt-logger-level==0.4.13
rqt-moveit==0.5.13
rqt-msg==0.4.12
rqt-nav-view==0.5.8
rqt-plot==0.4.16
rqt-pose-view==0.5.13
rqt-publisher==0.4.12
rqt-py-common==0.5.5
rqt-py-console==0.4.12
rqt-reconfigure==0.5.7
rqt-robot-dashboard==0.5.8
rqt-robot-monitor==0.5.15
rqt-robot-steering==0.5.14
rqt-runtime-monitor==0.5.10
rqt-rviz==0.7.2
rqt-service-caller==0.4.12
rqt-shell==0.4.13
rqt-srv==0.4.11
rqt-tf-tree==0.6.5
rqt-top==0.4.11
rqt-topic==0.4.15
rqt-web==0.4.11
rsa==4.9.1
rviz==1.14.26
scikit-image==0.21.0
scipy==1.10.1
seaborn==0.13.0
SecretStorage==2.3.1
sensor-msgs==1.13.2
service-identity==18.1.0
simplejson==3.16.0
sip==4.19.21
six==1.14.0
sklearn==0.0.post11
smach==2.5.3
smach-ros==2.5.3
smclib==1.8.7
smmap==5.0.1
snowballstemmer==2.0.0
solve==0.0.0
soupsieve==2.5
systemd-python==234
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.1
tensorflow-cpu-aws==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.35.0
tensorrt==8.4.1.5
termcolor==2.4.0
tf==1.13.4
tf-conversions==1.13.4
tf2-geometry-msgs==0.7.10
tf2-kdl==0.7.10
tf2-py==0.7.10
tf2-ros==0.7.10
thop==0.1.1.post2209072238
tifffile==2023.7.10
toml==0.10.0
topic-tools==1.17.3
torch @ file:///tmp/tmp.qJJc6SAEVw/torch-1.11.0-cp38-cp38-linux_aarch64.whl#sha256=f531a88331d233dfe75b550fa021f221c393c040122867a8398ea42c0d13b991
torchvision==0.12.0a0+9b5a3fe
tqdm==4.67.1
Twisted==18.9.0
txaio==2.10.0
typing_extensions==4.5.0
tzdata==2023.3
u-msgpack-python==2.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
uff==0.6.9
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==1.25.8
urwid==2.0.1
vcstools==0.1.42
wadllib==1.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==3.0.6
whitebox==2.3.1
wifite==2.5.2
wrapt==1.11.2
wsaccel==0.6.2
wstool==0.1.18
xacro==1.14.20
xkit==0.0.0
yolox==0.3.0
zipp==3.20.2
zope.interface==4.7.1
